feat(bf-w3): ONI T7 flag widen (registry+catalog bigint) + close-out artifact refreshes
- silver_noaa_oni: 4 ENSO flags tinyint->bigint (generator curation; the B3 canonical
  wrote the INV-2 int64 target; Glue applied via the F012 migrator, lease+backup, user gate)
- KNOWN_STALENESS emptied: every entry's catch-up wave closed with a certificate
  (BF-W1 weather trio, BF-W2 ESR vintages + WASDE); F082 alarms own freshness now
- hand DDLs synced (wasde 29-col post-F036, oni bigint, conab footer-refresh)

+++ sql/athena/ddl/silver_noaa_oni.sql
@@ -10,10 +10,10 @@ CREATE EXTERNAL TABLE IF NOT EXISTS silver_noaa_oni (
     oni_lag6               double,
     oni_lag9               double,
     oni_lag12              double,
-    el_nino_flag           tinyint,
-    la_nina_flag           tinyint,
-    la_nina_brazil_flag    tinyint,
-    argentina_la_nina_flag tinyint,
+    el_nino_flag           bigint,
+    la_nina_flag           bigint,
+    la_nina_brazil_flag    bigint,
+    argentina_la_nina_flag bigint,
     source                 string
 )
 STORED AS PARQUET
